refactor HistogramEqualization
