+++ tests/ImageSharp.Tests/Processing/Processors/Dithering/DitherTests.cs
@@ -50,8 +50,7 @@ namespace SixLabors.ImageSharp.Tests.Processing.Processors.Binarization
         /// The output is visually correct old 32bit runtime,
         /// but it is very different because of floating point inaccuracies.
         /// </summary>
-        private static readonly bool SkipAllDitherTests =
-            !TestEnvironment.Is64BitProcess && string.IsNullOrEmpty(TestEnvironment.NetCoreVersion);
+        private static readonly bool SkipAllDitherTests = false;
 
         [Theory]
         [WithFile(TestImages.Png.CalliphoraPartial, PixelTypes.Rgba32)]
